requirements.txt
python-dotenv
pandas
numpy 
matplotlib 
seaborn 
scikit-learn 
xgboost
mlflow
pymongo
certifi
pymongo[srv] == 3.6
pyaml
dill
xgboost
###-e .